fix dtype and update docs

+++ paddlemix/examples/ppdocbee/ppdocbee_infer.py
@@ -37,7 +37,7 @@ def main(args):
         logger.warning("bfloat16 is not supported on your device,change to float32")
         compute_dtype = "float32"
 
-    model = Qwen2VLForConditionalGeneration.from_pretrained(args.model_path, dtype="bfloat16")
+    model = Qwen2VLForConditionalGeneration.from_pretrained(args.model_path, dtype=compute_dtype)
 
     image_processor = Qwen2VLImageProcessor()
     tokenizer = MIXQwen2Tokenizer.from_pretrained(args.model_path)
